+++ api/rule/action/types.js
@@ -1,0 +1,5 @@
+const { UpdateAssigneeId } = require('./updateAssigneeId')
+
+module.exports = {
+  updateAssigneeId: (value) => new UpdateAssigneeId(value),
+}
