+++ api/rule/condition/index.js
@@ -1,75 +1,6 @@
 const _ = require('lodash')
 
-const { TICKET_STATUS } = require('../../lib/common')
-
-class UpdateTypeIs {
-  constructor(value) {
-    const values = ['create', 'update']
-    if (!values.includes(value)) {
-      throw new Error('Value must be one of ' + values.join(','))
-    }
-    this.value = value
-  }
-
-  test(ctx) {
-    return ctx.updateType === this.value
-  }
-}
-
-class StatusIs {
-  constructor(value) {
-    if (!(value in _.invert(TICKET_STATUS))) {
-      throw new Error('Invalid status')
-    }
-    this.value = value
-  }
-
-  test(ctx) {
-    return ctx.getTicketStatus() === this.value
-  }
-}
-
-class StatusIsNot extends StatusIs {
-  test(ctx) {
-    return !super.test(ctx)
-  }
-}
-
-class AssigneeIdIs {
-  constructor(value) {
-    if (typeof value !== 'string') {
-      throw new Error('Assignee ID must be a string')
-    }
-    this.value = value
-  }
-
-  test(ctx) {
-    if (this.value === '') {
-      return ctx.getTicketAssigneeId() === undefined
-    }
-    return ctx.getTicketAssigneeId() === this.value
-  }
-}
-
-class AssigneeIdIsNot extends AssigneeIdIs {
-  test(ctx) {
-    return !super.test(ctx)
-  }
-}
-
-const conditionFields = {
-  updateType: {
-    is: (value) => new UpdateTypeIs(value),
-  },
-  status: {
-    is: (value) => new StatusIs(value),
-    isNot: (value) => new StatusIsNot(value),
-  },
-  assigneeId: {
-    is: (value) => new AssigneeIdIs(value),
-    isNot: (value) => new AssigneeIdIsNot(value),
-  },
-}
+const conditionFields = require('./fields')
 
 class Condition {
   constructor(data, fields = conditionFields) {
